Isolate erc20 from benchmarks to reuse in tests
