+++ src/benches/eth_erc20.rs
@@ -1,10 +1,8 @@
-use crate::prelude::{Address, U256};
-use crate::transaction::EthTransaction;
+use crate::prelude::U256;
 use criterion::{BatchSize, BenchmarkId, Criterion};
 use secp256k1::SecretKey;
-use std::path::{Path, PathBuf};
 
-use crate::test_utils::solidity;
+use crate::test_utils::erc20::{ERC20Constructor, ERC20};
 use crate::test_utils::{address_from_secret_key, deploy_evm, sign_transaction, SUBMIT};
 
 const INITIAL_BALANCE: u64 = 1000;
@@ -110,110 +108,4 @@ pub(crate) fn eth_erc20_benchmark(c: &mut Criterion) {
     });
 
     group.finish();
-}
-
-struct ERC20Constructor(solidity::ContractConstructor);
-
-struct ERC20(solidity::DeployedContract);
-
-impl From<ERC20Constructor> for solidity::ContractConstructor {
-    fn from(c: ERC20Constructor) -> Self {
-        c.0
-    }
-}
-
-impl ERC20Constructor {
-    fn load() -> Self {
-        Self(solidity::ContractConstructor::compile_from_source(
-            Self::download_solidity_sources(),
-            Self::solidity_artifacts_path(),
-            "token/ERC20/presets/ERC20PresetMinterPauser.sol",
-            "ERC20PresetMinterPauser",
-        ))
-    }
-
-    fn deploy(&self, name: &str, symbol: &str, nonce: U256) -> EthTransaction {
-        let data = self
-            .0
-            .abi
-            .constructor()
-            .unwrap()
-            .encode_input(
-                self.0.code.clone(),
-                &[
-                    ethabi::Token::String(name.to_string()),
-                    ethabi::Token::String(symbol.to_string()),
-                ],
-            )
-            .unwrap();
-        EthTransaction {
-            nonce,
-            gas_price: Default::default(),
-            gas: Default::default(),
-            to: None,
-            value: Default::default(),
-            data,
-        }
-    }
-
-    fn download_solidity_sources() -> PathBuf {
-        let sources_dir = Path::new("target").join("openzeppelin-contracts");
-        let contracts_dir = sources_dir.join("contracts");
-        if contracts_dir.exists() {
-            contracts_dir
-        } else {
-            let url = "https://github.com/OpenZeppelin/openzeppelin-contracts";
-            let repo = git2::Repository::clone(url, sources_dir).unwrap();
-            // repo.path() gives the path of the .git directory, so we need to use the parent
-            repo.path().parent().unwrap().join("contracts")
-        }
-    }
-
-    fn solidity_artifacts_path() -> PathBuf {
-        Path::new("target").join("solidity_build")
-    }
-}
-
-impl ERC20 {
-    fn mint(&self, recipient: Address, amount: U256, nonce: U256) -> EthTransaction {
-        let data = self
-            .0
-            .abi
-            .function("mint")
-            .unwrap()
-            .encode_input(&[
-                ethabi::Token::Address(recipient),
-                ethabi::Token::Uint(amount),
-            ])
-            .unwrap();
-        EthTransaction {
-            nonce,
-            gas_price: Default::default(),
-            gas: Default::default(),
-            to: Some(self.0.address),
-            value: Default::default(),
-            data,
-        }
-    }
-
-    fn transfer(&self, recipient: Address, amount: U256, nonce: U256) -> EthTransaction {
-        let data = self
-            .0
-            .abi
-            .function("transfer")
-            .unwrap()
-            .encode_input(&[
-                ethabi::Token::Address(recipient),
-                ethabi::Token::Uint(amount),
-            ])
-            .unwrap();
-        EthTransaction {
-            nonce,
-            gas_price: Default::default(),
-            gas: Default::default(),
-            to: Some(self.0.address),
-            value: Default::default(),
-            data,
-        }
-    }
 }
